+++ src/egon/data/airflow/dags/pipeline.py
@@ -3,28 +3,34 @@ import os
 from airflow.operators.postgres_operator import PostgresOperator
 from airflow.operators.python_operator import PythonOperator
 from airflow.utils.dates import days_ago
-import airflow
 import importlib_resources as resources
 
 from egon.data.airflow.tasks import initdb
+from egon.data.processing.zensus_vg250 import (
+    zensus_population_inside_germany as zensus_vg250,
+)
+import airflow
 import egon.data.importing.demandregio as import_dr
+import egon.data.importing.demandregio.install_disaggregator as install_dr
 import egon.data.importing.etrago as etrago
+import egon.data.importing.heat_demand_data as import_hd
 import egon.data.importing.mastr as mastr
 import egon.data.importing.nep_input_data as nep_input
 import egon.data.importing.openstreetmap as import_osm
+import egon.data.importing.re_potential_areas as re_potential_areas
 import egon.data.importing.vg250 as import_vg250
+import egon.data.processing.demandregio as process_dr
 import egon.data.processing.openstreetmap as process_osm
 import egon.data.importing.zensus as import_zs
 import egon.data.processing.zensus as process_zs
+import egon.data.processing.osmtgmod as osmtgmod
 import egon.data.processing.power_plants as power_plants
-import egon.data.importing.nep_input_data as nep_input
-import egon.data.importing.etrago as etrago
-import egon.data.importing.mastr as mastr
 import egon.data.processing.substation as substation
 import egon.data.processing.zensus_vg250.zensus_population_inside_germany as zensus_vg250
-import egon.data.importing.re_potential_areas as re_potential_areas
-import egon.data.importing.heat_demand_data as import_hd
-import egon.data.processing.osmtgmod as osmtgmod
+import egon.data.processing.mv_grid_districts as mvgd
+import egon.data.importing.scenarios as import_scenarios
+import egon.data.importing.industrial_sites as industrial_sites
+import egon.data.processing.loadarea as loadarea
 from egon.data import db
 
 
@@ -153,40 +159,114 @@ with airflow.DAG(
     ] >> map_zensus_vg250 >> zensus_inside_ger >> zensus_inside_ger_metadata
     zensus_inside_ger >> vg250_population >> vg250_population_metadata
 
-    # DemandRegio data import
-    demandregio_import = PythonOperator(
-        task_id="import-demandregio",
-        python_callable=import_dr.insert_data,
+    # Scenario table
+    scenario_input_tables = PythonOperator(
+        task_id="create-scenario-parameters-table",
+        python_callable=import_scenarios.create_table
     )
-    vg250_clean_and_prepare >> demandregio_import
+
+    scenario_input_import = PythonOperator(
+        task_id="import-scenario-parameters",
+        python_callable=import_scenarios.insert_scenarios
+    )
+    setup >> scenario_input_tables >> scenario_input_import
+
+    # DemandRegio data import
+    demandregio_tables = PythonOperator(
+        task_id="demandregio-tables",
+        python_callable=import_dr.create_tables,
+    )
+
+    scenario_input_tables >> demandregio_tables
+
+
+    demandregio_installation = PythonOperator(
+        task_id="demandregio-installation",
+        python_callable=install_dr.clone_and_install,
+    )
+
+    setup >> demandregio_installation
+
+    demandregio_society = PythonOperator(
+        task_id="demandregio-society",
+        python_callable=import_dr.insert_society_data,
+    )
+
+    demandregio_installation >> demandregio_society
+    vg250_clean_and_prepare >> demandregio_society
+    demandregio_tables >> demandregio_society
+    scenario_input_import >> demandregio_society
+
+    demandregio_demand_households = PythonOperator(
+        task_id="demandregio-household-demands",
+        python_callable=import_dr.insert_household_demand,
+    )
+
+    demandregio_installation >> demandregio_demand_households
+    vg250_clean_and_prepare >> demandregio_demand_households
+    demandregio_tables >> demandregio_demand_households
+    scenario_input_import >> demandregio_demand_households
+
+    demandregio_demand_cts_ind = PythonOperator(
+        task_id="demandregio-cts-industry-demands",
+        python_callable=import_dr.insert_cts_ind_demands,
+    )
+
+    demandregio_installation >> demandregio_demand_cts_ind
+    vg250_clean_and_prepare >> demandregio_demand_cts_ind
+    demandregio_tables >> demandregio_demand_cts_ind
+    scenario_input_import >> demandregio_demand_cts_ind
 
     # Society prognosis
     prognosis_tables = PythonOperator(
         task_id="create-prognosis-tables",
-        python_callable=process_zs.create_tables
+        python_callable=process_zs.create_tables,
     )
 
     setup >> prognosis_tables
 
     population_prognosis = PythonOperator(
         task_id="zensus-population-prognosis",
-        python_callable=process_zs.population_prognosis_to_zensus
+        python_callable=process_zs.population_prognosis_to_zensus,
     )
 
     prognosis_tables >> population_prognosis
     map_zensus_vg250 >> population_prognosis
-    demandregio_import >> population_prognosis
+    demandregio_society >> population_prognosis
     population_import >> population_prognosis
 
     household_prognosis = PythonOperator(
         task_id="zensus-household-prognosis",
-        python_callable=process_zs.household_prognosis_to_zensus
+        python_callable=process_zs.household_prognosis_to_zensus,
     )
     prognosis_tables >> household_prognosis
     map_zensus_vg250 >> household_prognosis
-    demandregio_import >> household_prognosis
+    demandregio_society >> household_prognosis
     zensus_misc_import >> household_prognosis
 
+
+    # Distribute electrical demands to zensus cells
+    processed_dr_tables = PythonOperator(
+        task_id="create-demand-tables",
+        python_callable=process_dr.create_tables,
+    )
+
+    elec_household_demands_zensus = PythonOperator(
+        task_id="electrical-household-demands-zensus",
+        python_callable=process_dr.distribute_household_demands,
+    )
+
+    zensus_tables >> processed_dr_tables >> elec_household_demands_zensus
+    population_prognosis >> elec_household_demands_zensus
+    demandregio_demand_households >> elec_household_demands_zensus
+    map_zensus_vg250 >> elec_household_demands_zensus
+
+    # Power plant setup
+    power_plant_tables = PythonOperator(
+        task_id="create-power-plant-tables",
+        python_callable=power_plants.create_tables,
+    )
+    setup >> power_plant_tables
 
     # NEP data import
     create_tables = PythonOperator(
@@ -242,32 +322,47 @@ with airflow.DAG(
         autocommit=True,
     )
 
-    create_voronoi = PythonOperator(
-        task_id="create_voronoi",
-        python_callable=substation.create_voronoi
-    )
-    osm_add_metadata  >> substation_tables >> substation_functions
-    substation_functions >> hvmv_substation_extraction >> create_voronoi
-    substation_functions >> ehv_substation_extraction >> create_voronoi
+    osm_add_metadata >> substation_tables >> substation_functions
+    substation_functions >> hvmv_substation_extraction
+    substation_functions >> ehv_substation_extraction
     vg250_clean_and_prepare >> hvmv_substation_extraction
     vg250_clean_and_prepare >> ehv_substation_extraction
 
     # osmTGmod ehv/hv grid model generation
     run_osmtgmod = PythonOperator(
-        task_id= "run_osmtgmod",
-        python_callable= osmtgmod.run_osmtgmod,
+        task_id="run_osmtgmod",
+        python_callable=osmtgmod.run_osmtgmod,
     )
-    
 
     osmtgmod_pypsa = PythonOperator(
         task_id="osmtgmod_pypsa",
-        python_callable = osmtgmod.osmtgmmod_to_pypsa,
+        python_callable=osmtgmod.osmtgmmod_to_pypsa,
     )
 
+    osmtgmod_substation = PostgresOperator(
+        task_id="osmtgmod_substation",
+        sql=resources.read_text(osmtgmod, "substation_otg.sql"),
+        postgres_conn_id="egon_data",
+        autocommit=True,
+    )
     ehv_substation_extraction >> run_osmtgmod
     hvmv_substation_extraction >> run_osmtgmod
     run_osmtgmod >> osmtgmod_pypsa
+    run_osmtgmod >> osmtgmod_substation
 
+    # MV grid districts
+    create_voronoi = PythonOperator(
+        task_id="create_voronoi",
+        python_callable=substation.create_voronoi
+    )
+    osmtgmod_substation >> create_voronoi
+
+    
+    define_mv_grid_districts = PythonOperator(
+        task_id="define_mv_grid_districts",
+        python_callable=mvgd.define_mv_grid_districts
+    )
+    create_voronoi >> define_mv_grid_districts
 
     # Import potential areas for wind onshore and ground-mounted PV
     download_re_potential_areas = PythonOperator(
@@ -276,11 +371,11 @@ with airflow.DAG(
     )
     create_re_potential_areas_tables = PythonOperator(
         task_id="create_re_potential_areas_tables",
-        python_callable=re_potential_areas.create_tables
+        python_callable=re_potential_areas.create_tables,
     )
     insert_re_potential_areas = PythonOperator(
         task_id="insert_re_potential_areas",
-        python_callable=re_potential_areas.insert_data
+        python_callable=re_potential_areas.insert_data,
     )
     setup >> download_re_potential_areas >> create_re_potential_areas_tables
     create_re_potential_areas_tables >> insert_re_potential_areas
@@ -288,21 +383,72 @@ with airflow.DAG(
     # Future heat demand calculation based on Peta5_0_1 data
     heat_demand_import = PythonOperator(
         task_id="import-heat-demand",
-        python_callable=import_hd.future_heat_demand_data_import
+        python_callable=import_hd.future_heat_demand_data_import,
     )
     vg250_clean_and_prepare >> heat_demand_import
     zensus_inside_ger_metadata >> heat_demand_import
+    scenario_input_import >> heat_demand_import
 
     # Power plant setup
     power_plant_tables = PythonOperator(
         task_id="create-power-plant-tables",
-        python_callable=power_plants.create_tables
+        python_callable=power_plants.create_tables,
     )
 
     power_plant_import = PythonOperator(
         task_id="import-power-plants",
-        python_callable=power_plants.insert_power_plants
+        python_callable=power_plants.insert_power_plants,
     )
+
     setup >> power_plant_tables >> power_plant_import
     nep_insert_data >> power_plant_import
     retrieve_mastr_data >> power_plant_import
+
+    # Import and merge data on industrial sites from different sources
+
+    industrial_sites_import = PythonOperator(
+        task_id="download-import-industrial-sites",
+        python_callable=industrial_sites.download_import_industrial_sites
+    )
+
+    industrial_sites_merge = PythonOperator(
+        task_id="merge-industrial-sites",
+        python_callable=industrial_sites.merge_inputs
+    )
+
+    industrial_sites_nuts = PythonOperator(
+        task_id="map-industrial-sites-nuts3",
+        python_callable=industrial_sites.map_nuts3
+    )
+    vg250_clean_and_prepare >> industrial_sites_import
+    industrial_sites_import >> industrial_sites_merge >> industrial_sites_nuts
+
+    # Distribute electrical CTS demands to zensus grid
+
+    elec_cts_demands_zensus = PythonOperator(
+        task_id="electrical-cts-demands-zensus",
+        python_callable=process_dr.distribute_cts_demands,
+    )
+
+    processed_dr_tables >> elec_cts_demands_zensus
+    heat_demand_import >> elec_cts_demands_zensus
+    demandregio_demand_cts_ind >> elec_cts_demands_zensus
+    map_zensus_vg250 >> elec_cts_demands_zensus
+
+    # Extract landuse areas from osm data set
+    create_landuse_table = PythonOperator(
+        task_id="create-landuse-table",
+        python_callable=loadarea.create_landuse_table
+    )
+
+    landuse_extraction = PostgresOperator(
+        task_id="extract-osm_landuse",
+        sql=resources.read_text(loadarea, "osm_landuse_extraction.sql"),
+        postgres_conn_id="egon_data",
+        autocommit=True,
+    )
+    setup >> create_landuse_table
+    create_landuse_table >> landuse_extraction
+    osm_add_metadata >> landuse_extraction
+    vg250_clean_and_prepare >> landuse_extraction
+