+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Added
 
 * OSM data import as done in open_ego
   `#1 <https://github.com/openego/eGon-data/issues/1>`_
+  which was updated to the latest long-term data set of the 2021-01-01 in
+  #223 <https://github.com/openego/eGon-data/issues/223>`_
 * Verwaltungsgebiete data import (vg250) more or less done as in open_ego
   `#3 <https://github.com/openego/eGon-data/issues/3>`_
 * Zensus population data import
@@ -65,7 +67,19 @@ Added
   `#127 <https://github.com/openego/eGon-data/issues/127>`_
 * Creation of the ehv/hv grid model with osmTGmod, see 
   `issue #4 <https://github.com/openego/eGon-data/issues/4>`_ and 
-  `PR #164 <https://github.com/openego/eGon-data/pull/164>`_  
+  `PR #164 <https://github.com/openego/eGon-data/pull/164>`_
+* Identification of medium-voltage grid districts
+  `#10 <https://github.com/openego/eGon-data/pull/10>`_  
+* Distribute electrical demands of households to zensus cells
+  `#181 <https://github.com/openego/eGon-data/issues/181>`_
+* Distribute electrical demands of cts to zensus cells
+  `#210 <https://github.com/openego/eGon-data/issues/210>`_
+* Include industrial sites' download, import and merge
+  `#117 <https://github.com/openego/eGon-data/issues/117>`_
+* Integrate scenario table with parameters for each sector 
+  `#177 <https://github.com/openego/eGon-data/issues/177>`_
+* Extract landuse areas from OSM 
+  `#214 <https://github.com/openego/eGon-data/issues/214>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 
@@ -98,6 +112,8 @@ Changed
   `#185 <https://github.com/openego/eGon-data/issues/185>`_
 * Update task zensus-inside-germany
   `#196 <https://github.com/openego/eGon-data/issues/196>`_
+* Update installation of demandregio's disaggregator
+  `#202 <https://github.com/openego/eGon-data/issues/202>`_
 
 Bug fixes
 ---------
@@ -105,3 +121,11 @@ Bug fixes
   `#157 <https://github.com/openego/eGon-data/issues/157>`_
 * Substation sequence
   `#171 <https://github.com/openego/eGon-data/issues/171>`_
+* Adjust names of demandregios nuts3 regions according to nuts version 2016
+  `#201 <https://github.com/openego/eGon-data/issues/201>`_
+* Delete zensus buildings, apartments and households in unpopulated cells
+  `#202 <https://github.com/openego/eGon-data/issues/202>`_
+* Fix input table of electrical-demands-zensus
+  `#217 <https://github.com/openego/eGon-data/issues/217>`_
+* Import heat demand raster files successively to fix import for dataset==Everything
+  `#204 <https://github.com/openego/eGon-data/issues/204>`_